store events and url_history per node
when printing for a specific node (ie: a specific unit test), we were
using filters and lambdas to extract the list of events/urls that match
that given node.

when we added this code we accidentally broke the de-duplication code
(`current_url != self._url_history[-1]` in `_add_event`). This was
comparing the current_url to the url_history, but the string in
url_history contained the node as well, so this check would never pass
and we'd fill `_url_history` up with duplicate entries, reducing its
usefulness.

To fix this, rather than changing that log, just change how we store the
events to store them in a dictionary per-node instead, so we never need
to do any filtering.