save screenshots to a single filename by default
on concourse, if functional tests retry, and both test runs fail, then
we'll end up with two separate failure screenshots. However, the
concourse resource that puts these on s3 fails because it expects
exactly one file[^1]. We could zip the screenshots and upload them all,
but this makes it more awkward for a dev to download, extract, and view
them rather than just clicking the link in concourse.

So for now, we can just splat over the first failure by having a static
filename, and be sure that we'll only upload one. If you wanted to see a
screenshot for the first failure, well, tough! sorry!

Because it's still useful to be able to save nice screenshots locally,
add a new command line option to save them using verbose names. This can
be triggered by running `pytest --unique-screenshot-filenames ...`

[^1]: https://github.com/concourse/s3-resource/issues/55